+++ src/Core/ArgsParser.py
@@ -465,8 +465,9 @@ def parse_arguments():
                              "Example: --client=synology --remove-duplicates-albums")
 
     PARSER.add_argument("-rDupAst", "--remove-duplicates-assets", action="store_true", default="",
-                        help="Remove duplicate assets with the same filename and size, preserving available metadata on one keeper.\n"
-                             "Requires '--client' and the required module selector '--duplicate-asset-keeper {oldest, newest}'.\n"
+                        help="Remove duplicate assets with the same filename and size, preserving all available metadata and people from duplicated assets on the selected keeper assetr.\n"
+                             "Requires '--client'.\n"
+                             "Also requires module selector '--duplicate-asset-keeper {oldest, newest}'.\n"
                              "The selector defaults to 'newest'. Google Photos cannot delete library media through its public API.\n"
                              "Example: --client=immich --remove-duplicates-assets --duplicate-asset-keeper newest")